+++ nip/utils/plotting/__init__.py
@@ -2,4 +2,4 @@
 
 from .decision_agreement import DecisionAgreementAnalysis, analyse_decision_agreement
 from .rollouts import get_last_timestep_mask
-from .decision_scale import get_decision_histogram
+from .decision_scale import get_decision_histogram, get_thresholded_performance